refactor: improve some design aspects in the drone system
